move remote layer protocl out of intproxy protocol
