+++ mirrord/remote/layer/protocol/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "mirrord-remote-layer-protocol"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+publish.workspace = true
+edition.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+bincode.workspace = true
